core: Implement a hard limit on recursion via subprocesses
If a chain of subdirectories exist, each of which are the most interesting
child of their parent directory, GnomeDesktopThumbnailFactory will
unavoidably spawn a series of gnome-directory-thumbnailer subprocesses. It’s
possible to encounter a directory loop (e.g. via a bind mount to a parent
directory), so this commit implements a hard limit on the recursion depth to
avoid infinite loops of subprocesses.

This is implemented using a GNOME_DIRECTORY_THUMBNAILER_RECURSION_LIMIT
environment variable which is decremented for each subprocess.